requirements.txt
aioschedule==0.5.2
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.1
APScheduler==3.9.1
async-lru==1.0.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
h11==0.12.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
idna==3.3
Mako==1.2.1
MarkupSafe==2.1.1
oauthlib==3.2.0
protobuf==4.21.4
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
Pyrogram==1.2.9
PySocks==1.7.1
python-dotenv==0.20.0
python-telegram-bot==20.0a2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
tornado==6.2
ttl-cache==1.6
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
